Move non-migratable Angular dependencies into directives folder  (#41638)
* Moved render_directive to directives folder

* Moved debounce to directives

* Moved bind to directives

* Moved watch multi into directives

* Moved listen to directives
