+++ src/legacy/core_plugins/timelion/public/app.js
@@ -39,7 +39,7 @@ require('ui/autoload/all');
 // TODO: remove ui imports completely (move to plugins)
 import 'ui/directives/input_focus';
 import 'ui/directives/saved_object_finder';
-import 'ui/listen';
+import 'ui/directives/listen';
 import 'ui/kbn_top_nav';
 import 'ui/saved_objects/ui/saved_object_save_as_checkbox';
