+++ src/legacy/ui/public/courier/courier.js
@@ -22,7 +22,7 @@ import _ from 'lodash';
 import { timefilter } from 'ui/timefilter';
 
 import '../es';
-import '../listen';
+import '../directives/listen';
 import '../index_patterns';
 import { uiModules } from '../modules';
 import { addFatalErrorCallback } from '../notify';
