+++ src/legacy/ui/public/directives/bind/__tests__/bind.js
