+++ src/legacy/ui/public/directives/bind/index.js
