+++ src/legacy/ui/public/directives/debounce/debounce.js
@@ -18,7 +18,7 @@
  */
 
 import _ from 'lodash';
-import { uiModules } from '../modules';
+import { uiModules } from '../../modules';
 // Debounce service, angularized version of lodash debounce
 // borrowed heavily from https://github.com/shahata/angular-debounce
