+++ src/legacy/ui/public/directives/listen/listen.js
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import { uiModules } from '../modules';
+import { uiModules } from '../../modules';
 
 uiModules.get('kibana')
   .run(function ($rootScope) {
