+++ src/legacy/ui/public/directives/render_directive/__tests__/render_directive.js
