+++ src/legacy/ui/public/directives/render_directive/index.js
