+++ src/legacy/ui/public/directives/watch_multi/__tests__/watch_multi.js
