+++ src/legacy/ui/public/directives/watch_multi/index.js
