+++ src/legacy/ui/public/directives/watch_multi/watch_multi.js
@@ -18,8 +18,8 @@
  */
 
 import _ from 'lodash';
-import { uiModules } from '../modules';
-import { callEach } from '../utils/function';
+import { uiModules } from '../../modules';
+import { callEach } from '../../utils/function';
 
 uiModules.get('kibana')
   .config(function ($provide) {
