+++ src/legacy/ui/public/kbn_top_nav/kbn_top_nav.js
@@ -56,7 +56,7 @@
 import _ from 'lodash';
 import angular from 'angular';
 import './timepicker';
-import '../watch_multi';
+import '../directives/watch_multi';
 import '../directives/input_focus';
 import { uiModules } from '../modules';
 import template from './kbn_top_nav.html';
