+++ x-pack/legacy/plugins/monitoring/public/views/no_data/controller.js
@@ -16,7 +16,7 @@ import { render, unmountComponentAtNode } from 'react-dom';
 import { NoData } from 'plugins/monitoring/components';
 import { timefilter } from 'ui/timefilter';
 import { I18nContext } from 'ui/i18n';
-import 'ui/listen';
+import 'ui/directives/listen';
 
 const REACT_NODE_ID_NO_DATA = 'noDataReact';
